web-app-requirements.txt
tensorflow
numpy
pillow
scipy
jupyter
flask
